AEROGEAR-2925 Update app icons and display name
Currently the display name of the app is its Bundle Identifier.
Along with this its icon wasn't defined, causing an unattrative
default icon to be shown.

This adds the AeroGear icon to the app. It also updates the display
name of the app to be something more readable.

It should be noted that in iOS 11 a change was introduced which
removes spaces in apps with names larger than 12 characters. This
means no spaces show up in the app name.