Show push registration error message
Currently when push fails to register we create a log message but
do not show the end user of the app directly. This is important to
do as push messages do not ever successfully register on the iOS
simulator so the users will be left with a blank screen if they go
to the page on a simulator.

This adds a message saying that the push registration failed with
a details section that shows the message provided with the push
registration error.